requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
colorama==0.4.3
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyparsing==2.4.7
pytest==5.4.2
six==1.14.0
wcwidth==0.1.9
